install.py

+++ install.py
@@ -1,6 +1,7 @@
 import os
 import subprocess
 import shutil
+from modules import shared
 
 
 extension_root = os.path.dirname(os.path.realpath(__file__))
@@ -8,7 +9,7 @@ package_json = os.path.join(extension_root, 'package.json')
 
 
 def install_localtunnel():
-    if not os.path.exists(package_json):
+    if shared.cmd_opts.localtunnel_comfyui and not os.path.exists(package_json):
         subprocess.run('npm install localtunnel', cwd=extension_root, shell=True)
